third_party/blink/web_tests/css3/flexbox/block-available-size.html

<!DOCTYPE html>
<style>
#flex {
  display: flex;
  background: red;
  width: 100px;
  height: 100px;
  align-items: start;
}

#item {
  width: 100px;
  background: green;
  height: -webkit-fill-available;
}
</style>
<p>Test passes if there is a filled green square and <strong>no red</strong>.</p>
<div id=flex>
  <div id=item data-expected-height=100></div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
checkLayout("#flex");
</script>
